// packages/ui/src/client/assethub/abi/Curation.ts
export type CurationAssetStruct = {
  hub: AddressLike;
  assetId: BigNumberish;
  order: BigNumberish;
};